+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.2
 
 - Enumerized Edition Direction
+- Fetch only one surah by number and edition(optional) with client.getSurah or client.getSurahByEdition
 
 ## 0.0.1